examples/worker-qwen3-vl/pyproject.toml
[project]
name = "worker-qwen3-vl"
version = "0.1.0"
description = "Simple CV worker that receives jobs and logs them"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "websockets>=12.0",
    "requests>=2.31.0",
    "transformers>=4.37.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pillow>=10.0.0",
    "qwen-vl-utils",
]
